chore: Create new folder structure, move some components
