+++ apps/editor.planx.uk/src/pages/FlowEditor/components/forms/AddComponentModal/AddComponentModal.stories.tsx
@@ -3,7 +3,7 @@ import type { Meta, StoryObj } from "@storybook/tanstack-react";
 
 import AddComponentModal, {
   AddComponentModalContent,
-} from "./AddComponentModal";
+} from "../AddComponentModal";
 
 const meta: Meta<typeof AddComponentModal> = {
   title: "Editor Components/Modal/AddComponentModal",
